# fix more performance issues in `flux resource`