--- a/src/bindings/python/flux/resource/status.py
+++ b/src/bindings/python/flux/resource/status.py
@@ -112,10 +112,14 @@ class ResourceStatus:
             ranks = IDset(drain_ranks)
             if include_ranks is not None:
                 ranks = ranks.intersect(include_ranks)
-            self.drained += ranks - self.allocated
-            self.draining += ranks - self.drained
+            self.drained += ranks
             info = DrainInfo(ranks, entry["timestamp"], entry["reason"])
             self.drain_info.append(info)
+
+        # create the set of draining ranks as the intersection of
+        #  drained and allocated
+        self.draining = self.drained & self.allocated
+        self.drained -= self.draining
 
         # available: all ranks not excluded or drained/draining
         self.avail = self.all - self.get_idset("exclude", "drained", "draining")